#9 Move Validity to the neutral dev.hardwood package
Validity is about to become shared read/write vocabulary: the writer accepts it
as a column's null mask, mirroring what the reader returns. Owning it in the
reader package would make the writer depend on the reader for a concept neither
owns, so relocate the interface to dev.hardwood and its implementations to
dev.hardwood.internal, updating references across the reader, tests, and
benchmarks. Pure relocation; behaviour is unchanged.
