+++ core/src/test/java/dev/hardwood/ColumnReaderLayerModelTest.java
@@ -16,7 +16,6 @@ import org.junit.jupiter.api.Test;
 import dev.hardwood.reader.ColumnReader;
 import dev.hardwood.reader.LayerKind;
 import dev.hardwood.reader.ParquetFileReader;
-import dev.hardwood.reader.Validity;
 
 import static org.assertj.core.api.Assertions.assertThat;
