+++ core/src/test/java/dev/hardwood/ColumnReadersTest.java
@@ -16,7 +16,6 @@ import org.junit.jupiter.api.Test;
 import dev.hardwood.reader.ColumnReader;
 import dev.hardwood.reader.ColumnReaders;
 import dev.hardwood.reader.ParquetFileReader;
-import dev.hardwood.reader.Validity;
 import dev.hardwood.schema.ColumnProjection;
 
 import static org.assertj.core.api.Assertions.assertThat;
