+++ core/src/test/java/dev/hardwood/NestedDictBatchBoundaryTest.java
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.Test;
 
 import dev.hardwood.reader.ColumnReader;
 import dev.hardwood.reader.ParquetFileReader;
-import dev.hardwood.reader.Validity;
 
 import static org.assertj.core.api.Assertions.assertThat;
