+++ core/src/test/java/dev/hardwood/ParquetReaderTest.java
@@ -22,7 +22,6 @@ import dev.hardwood.reader.FilterPredicate;
 import dev.hardwood.reader.LayerKind;
 import dev.hardwood.reader.ParquetFileReader;
 import dev.hardwood.reader.RowReader;
-import dev.hardwood.reader.Validity;
 import dev.hardwood.schema.ColumnProjection;
 import dev.hardwood.schema.ColumnSchema;
 import dev.hardwood.schema.FileSchema;
